% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startSubVis.R
\name{startSubVis}
\alias{startSubVis}
\title{Starts the SubVis app}
\usage{
startSubVis()
}
\description{
Author:  Scott Barlowe
Date:  June 15, 2016
}
\examples{

\donttest{
startSubVis()
}

}

